+++ Runtime/AssemblyInternals.cs
@@ -1,4 +1,6 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("Unity.Services.PushNotifications.Tests")]
+[assembly: InternalsVisibleTo("Unity.Services.PushNotifications.EditorTests")]
 [assembly: InternalsVisibleTo("Unity.Services.PushNotifications.Editor")]
+[assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]